Ojasritu-Wellness-ojasritu/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
stripe>=5.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
social-auth-app-django>=5.0.0
django-allauth>=0.57.0
django-browser-reload>=1.12.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
openai>=1.0.0
google-auth>=2.0.0
requests>=2.28.0
